+++ app/src/androidTest/java/com/example/tastezzip/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.tastezip
+package com.example.tastezzip
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
