+++ app/src/main/java/com/example/tastezzip/ui/component/CustomIconButton.kt
@@ -1,8 +1,7 @@
-package com.example.tastezip.ui.component
+package com.example.tastezzip.ui.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.material.Text
 import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -11,7 +10,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.unit.sp
-import com.example.tastezip.R
+import com.example.tastezzip.R
 
 @Composable
 fun CustomIconButton(
